Create V0.3
Mode change more stable. Only "off" and "heat" modes supported. After using the "off" command, use the "auto" command to switch the TRV back on. Manual selection (on the TRV) of modes ECO and AUTO are permitted - these will show as "heat" on the dashboard, and the temperatures associated with these modes will be current. e.g. if ECO mode is selected on TRV with a temp of 5C, Hubitat will be in mode "heat", and the setpoint will be 5C. Temperature adjustment (0-5c) is available, but this is not fully tested.